+++ jdi-light-angular-tests/src/test/java/io/github/epam/angular/tests/elements/complex/datepicker/InputChangeEventsDatepickerTests.java
@@ -16,7 +16,7 @@ public class InputChangeEventsDatepickerTests extends TestsDatepickerBase {
 
     @BeforeMethod(alwaysRun = true)
     public void before() {
-        angularPage.open();
+        angularPage.shouldBeOpened();
         inputChangeEventsDatepicker.show();
         inputChangeEventsDatepicker.clear();
     }