+++ jdi-light-angular-tests/src/test/java/io/github/epam/angular/tests/unit/BottomSheetUnitTests.java
@@ -1,12 +1,15 @@
 package io.github.epam.angular.tests.unit;
 
 import io.github.epam.TestsInit;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
 import java.util.List;
 
+import static com.epam.jdi.light.elements.composite.WebPage.reload;
 import static io.github.com.StaticSite.angularPage;
 import static io.github.com.pages.AngularPage.bottomSheet;
 import static io.github.epam.site.steps.States.shouldBeLoggedIn;
@@ -14,12 +17,20 @@ import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
 public class BottomSheetUnitTests extends TestsInit {
-    private static final List<String> BOTTOM_SHEET_VALUES = Arrays.asList("Google Keep", "Google Docs", "Google Plus", "Google Hangouts");
+    private static final List<String> BOTTOM_SHEET_VALUES =
+            Arrays.asList("Google Keep", "Google Docs", "Google Plus", "Google Hangouts");
 
     @BeforeClass(alwaysRun = true)
     public void before() {
         shouldBeLoggedIn();
-        angularPage.open();
+        angularPage.shouldBeOpened();
+        bottomSheet.show();
+    }
+
+    @BeforeMethod
+    // required to reload before each test
+    public void reloadPage() {
+        reload();
         bottomSheet.show();
     }
 
@@ -42,4 +53,8 @@ public class BottomSheetUnitTests extends TestsInit {
         assertEquals(bottomSheet.values(), BOTTOM_SHEET_VALUES);
     }
 
+    @AfterClass
+    public void tearDownClass() {
+        reload();
+    }
 }