+++ jdi-light-mobile-tests/src/main/java/nativeapp/contacts/CreateNewContact.java
@@ -2,7 +2,6 @@ package nativeapp.contacts;
 
 import com.epam.jdi.light.elements.pageobjects.annotations.FindBy;
 import com.epam.jdi.light.mobile.elements.common.app.ios.TextField;
-
 import com.epam.jdi.light.ui.html.elements.complex.DataListOptions;
 
 public class CreateNewContact {