+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/MobileFileManagerInit.java
@@ -9,7 +9,9 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintStream;
 import java.nio.file.Files;
 
 import static com.epam.jdi.light.mobile.elements.init.PageFactory.initMobile;