+++ jdi-light-mobile/src/main/java/com/epam/jdi/light/mobile/settings/MobileSettings.java
@@ -8,12 +8,9 @@ import com.epam.jdi.light.mobile.elements.base.MobileUIElement;
 import com.epam.jdi.light.settings.JDISettings;
 import com.epam.jdi.light.settings.WebSettings;
 
-import static com.epam.jdi.light.driver.get.RemoteDriverInfo.SAUCE_LABS;
-import static com.epam.jdi.light.driver.get.RemoteDriverInfo.appium;
-import static com.epam.jdi.light.driver.get.RemoteDriverInfo.browserstack;
-import static com.epam.jdi.light.elements.base.JdiSettings.DEFAULT_CONTEXT;
-
+import static com.epam.jdi.light.driver.get.RemoteDriverInfo.*;
 import static com.epam.jdi.light.driver.sauce.SauceSettings.sauceCapabilities;
+import static com.epam.jdi.light.elements.base.JdiSettings.DEFAULT_CONTEXT;
 import static com.epam.jdi.light.elements.init.PageFactory.STOP_INIT_CLASSES;
 import static com.epam.jdi.light.mobile.driver.MobileDriverData.CAPABILITIES_FOR_ANDROID;
 import static com.epam.jdi.light.mobile.driver.MobileDriverData.CAPABILITIES_FOR_IOS;