+++ jdi-light/src/main/java/com/epam/jdi/light/elements/composite/WebPage.java
@@ -9,6 +9,7 @@ import com.epam.jdi.light.elements.pageobjects.annotations.Title;
 import com.epam.jdi.light.elements.pageobjects.annotations.Url;
 import com.epam.jdi.tools.CacheValue;
 import com.epam.jdi.tools.Safe;
+import com.epam.jdi.tools.Timer;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.logging.LogEntry;
@@ -20,6 +21,7 @@ import java.text.MessageFormat;
 import java.util.List;
 import java.util.function.Supplier;
 
+import static com.epam.jdi.light.actions.ActionProcessor.isTop;
 import static com.epam.jdi.light.common.CheckTypes.*;
 import static com.epam.jdi.light.common.Exceptions.exception;
 import static com.epam.jdi.light.common.OutputTemplates.*;
@@ -210,6 +212,7 @@ public class WebPage extends DriverBase implements PageObject {
         CacheValue.reset();
         driver().navigate().to(url);
         getWindows();
+        isTop.set(true);
         setCurrentPage(this);
     }
     public void open(Object... params) {
@@ -236,7 +239,7 @@ public class WebPage extends DriverBase implements PageObject {
     @JDIAction("Check that '{name}' is opened (url {checkUrlType} '{checkUrl}'; title {checkTitleType} '{title}')")
     public void checkOpened() {
         if (noRunDrivers())
-            throw exception("Page '%s' is not opened: Driver is not run", toString());
+            throw exception("Page '%s' is not opened: Driver is not run: ", toString());
         String result = Switch(checkUrlType).get(
             Value(NONE, ""),
             Value(EQUALS, t -> !url().check() ? "Url '%s' doesn't equal to '%s'" : ""),
@@ -253,9 +256,18 @@ public class WebPage extends DriverBase implements PageObject {
         );
         if (isNotBlank(result))
             throw exception("Page '%s' is not opened: %s", getName(), format(result, driver().getTitle(), title));
-        setCurrentPage(this);
         if (VISUAL_PAGE_STRATEGY == CHECK_PAGE)
             visualWindowCheck();
+        isTop.set(true);
+        setCurrentPage(this);
+    }
+    public void checkIsNotChanged() {
+        if (noRunDrivers())
+            throw exception("Driver is not run: ", toString());
+        boolean result = new Timer(TIMEOUTS.page.get() * 1000L).getResult(() -> !isOpened());
+        if (!result) {
+            throw exception("New page opened: %s", getUrl());
+        }
     }
     public boolean isOnPage(String url) {
         switch (checkUrlType) {
@@ -298,8 +310,10 @@ public class WebPage extends DriverBase implements PageObject {
             Value(CONTAINS, t -> title().contains()),
             Else(false)
         );
-        if (result)
+        if (result) {
+            isTop.set(true);
             setCurrentPage(this);
+        }
         return result;
     }