+++ jdi-light/src/main/java/com/epam/jdi/light/settings/WebSettings.java
@@ -189,6 +189,8 @@ public class WebSettings {
             fillAction(WebSettings::setSearchStrategy, "element.search.strategy");
             fillAction(p -> DRIVER.screenSize.read(p), "browser.size");
             fillAction(p -> DRIVER.pageLoadStrategy = getPageLoadStrategy(p), "page.load.strategy");
+            fillAction(p -> DRIVER.gitHubTokenName = p, "gitHubTokenName");
+            fillAction(p -> DRIVER.gitHubTokenSecret = p, "gitHubTokenSecret");
             fillAction(p -> PAGE.checkPageOpen = parse(p), "page.check.after.open");
             fillAction(SoftAssert::setAssertType, "assert.type");
             fillAction(p -> ELEMENT.clickType = getEnumValueByName(ElementArea.class, p, CENTER), "click.type");
@@ -289,15 +291,23 @@ public class WebSettings {
         String path = "";
         try {
             path = System.getProperty(property, getProperty(property));
-        } catch (Exception ignore) { }
+        } catch (Exception ignore) {
+            logger.trace("Error on read property %s. %s", property, ignore);
+        }
         if (isEmpty(path))
             path = defaultPath;
         Properties properties = new PropReader(path).getProperties();
-        if (properties.isEmpty())
+        if (properties.isEmpty()) {
+            logger.trace("There is no properties in %s", property);
+            File p = new File("."+path);
+            logger.trace("Abs prop path %s, Properties file exists? = %s", p.getAbsolutePath(), p.exists());
             return;
+        }
         try {
             setCapabilities.execute(properties);
-        } catch (Exception ignore) { }
+        } catch (Exception ignore) {
+            logger.trace("Error set properties %s. %s", property, ignore);
+        }
     }
 
     private static void setSearchStrategy(String p) {
@@ -354,7 +364,7 @@ public class WebSettings {
         Properties properties = new Properties();
         try {
             properties.load(new FileInputStream(propertyFile));
-            System.out.println("Property file found: " + propertyFile.getAbsolutePath());
+            logger.info("Property file found: %s", propertyFile.getAbsolutePath());
         } catch (IOException ex) {
             throw exception("Couldn't load properties for CI Server" + path);
         }